Commit Fully Functioning PlayerController
The playercontroller finally works, and I think it has accounted for basically everything. It was a lot of work, but it is done. Can now start working on some progress bars for the special timers, and for the dash. If that is done, then go on to enemy AI and all different kinds of enemies. Awesome!